+++ src/rex.setup/test/fixtures/basic_shrinkwrap/static/js/lib/index.js
@@ -1,0 +1,2 @@
+console.log(require('json-loader!esprima/package.json').version);
+console.log("PASSED!");